+++ frontend/src/router.js
@@ -9,6 +9,7 @@ import AideNoteDetailView from "./views/AideNoteDetailView.js";
 import AideNoteFormView from "./views/AideNoteFormView.js";
 import AideNoteListView from "./views/AideNoteListView.js";
 import AideNotePrintView from "./views/AideNotePrintView.js";
+import BrandingSettingsView from "./views/BrandingSettingsView.js";
 import NurseNotePrintView from "./views/NurseNotePrintView.js";
 import PatientDetailView from "./views/PatientDetailView.js";
 import PatientFormView from "./views/PatientFormView.js";
@@ -178,6 +179,12 @@ const routes = [
     name: "nurse-note-print",
     component: NurseNotePrintView,
     props: true,
+  },
+  {
+    path: "/settings/branding",
+    name: "branding-settings",
+    component: BrandingSettingsView,
+    meta: { roles: ["admin", "manager"] },
   },
 ];